requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
distlib==0.3.9
dj-database-url==3.0.0
Django==4.2.16
django-active-link==0.2.2
django-choices==2.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-formtools==2.5.1
django-json-widget==2.0.1
djangorestframework==3.16.1
dotenv==0.9.9
filelock==3.18.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
kombu==5.5.4
packaging==23.2
pillow==10.4.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psycopg==3.2.3
psycopg2==2.9.9
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-env==1.0.0
redis==5.2.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
whitenoise==6.7.0
